+++ src/main/kotlin/icu/windea/pls/core/diff/actions/CompareFilesAction.kt
@@ -31,30 +31,31 @@ import javax.swing.*
 /**
  * 将当前文件与包括当前文件的只读副本在内的相同路径的文件进行DIFF。如果是本地化文件的话也忽略路径中的语言区域。
  *
- * * 当当前文件是模组或游戏文件时显示并启用。
+ * * 当当前文件（或者通过视图等选中的文件）是模组或游戏文件时显示并启用。
  * * 忽略直接位于游戏或模组入口目录下的文件。
  * * 可以用于比较二进制文件。（如DDS图片）
  * * TODO 按照覆盖顺序进行排序。
  */
-@Suppress("ComponentNotRegistered", "UNUSED_VARIABLE", "DEPRECATION")
+@Suppress("ComponentNotRegistered", "DEPRECATION")
 class CompareFilesAction : ParadoxShowDiffAction() {
     private fun findFile(e: AnActionEvent): VirtualFile? {
-        return e.getData(CommonDataKeys.VIRTUAL_FILE)
-            ?: e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)?.singleOrNull()
+        val file = e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)?.singleOrNull()
+            ?: e.getData(CommonDataKeys.VIRTUAL_FILE)
+            ?: return null
+        if(file.isDirectory) return null
+        val fileInfo = file.fileInfo ?: return null
+        if(fileInfo.entryPath.length <= 1) return null //忽略直接位于游戏或模组入口目录下的文件
+        //val gameType = fileInfo.rootInfo.gameType
+        //val path = fileInfo.path.path
+        return file
     }
     
     override fun update(e: AnActionEvent) {
         val presentation = e.presentation
-        presentation.isVisible = false
-        presentation.isEnabled = false
+        presentation.isEnabledAndVisible = false
         val project = e.project ?: return
-        val file = findFile(e) ?: return
-        if(file.isDirectory) return
-        val fileInfo = file.fileInfo ?: return
-        if(fileInfo.entryPath.length <= 1) return //忽略直接位于游戏或模组入口目录下的文件
+        val file = findFile(e)
         presentation.isVisible = true
-        //val gameType = fileInfo.rootInfo.gameType
-        //val path = fileInfo.path.path
         //val selector = fileSelector(project, file)
         //val multiple = ParadoxFilePathSearch.search(path, project, selector).hasMultipleResults()
         //if(!multiple) return //忽略不存在重载/被重载的情况 - 出于性能原因，目前不在update方法中判断